+++ golang/lib/client/do_nothing_client.go
@@ -19,8 +19,8 @@ func (client *doNothingClient) TrackShouldSendMetricsUserElection(didUserAcceptS
 	return nil
 }
 
-func (client *doNothingClient) TrackCreateEnclave(enclaveId string) error {
-	logrus.Debugf("Do-nothing metrics client TrackCreateEnclave called with argument enclaveId '%v'; skipping sending event", enclaveId)
+func (client *doNothingClient) TrackCreateEnclave(enclaveId string, isSubnetworkingEnabled bool) error {
+	logrus.Debugf("Do-nothing metrics client TrackCreateEnclave called with argument enclaveId '%v', isSubnetworkingEnabled '%v'; skipping sending event", enclaveId, isSubnetworkingEnabled)
 	client.callback.Success()
 	return nil
 }