+++ golang/lib/client/metrics_client.go
@@ -2,7 +2,7 @@ package client
 
 type MetricsClient interface {
 	TrackShouldSendMetricsUserElection(didUserAcceptSendingMetrics bool) error
-	TrackCreateEnclave(enclaveId string) error
+	TrackCreateEnclave(enclaveId string, isSubnetworkingEnabled bool) error
 	TrackStopEnclave(enclaveId string) error
 	TrackDestroyEnclave(enclaveId string) error
 	TrackKurtosisRun(packageId string, isRemote bool, isDryRun bool, isScript bool) error